[+FEATURE] Extbase (MVC): Automatic target page determination
you can use the "pageUid" argument of the link.* and uri.* view helpers
to link to a different page. That is deprecated though as we won't have
the notion of "page uids" in v5. Instead the target page is now determined
automatically.
If the target page can't be determined because more than one active
plugin is capable of handling the action an exception will be thrown.
In that case you'll have to define the target page either by using the
pageUid argument or - preferably - by setting
plugin.tx_extensionname_pluginname.view.defaultPid to a fixed page uid.

Note: This feature still has to be documented!

Resolves: #9121