[+BUGFIX] Extbase: Some smaller tweaks and fixes [!!!][+FEATURE] Extbase (Persistence): Backport QueryResult from FLOW3
Now Query::execute() returns an instance of QueryResultInterface that allows it to modify the query before actually accessing the records that it retrieves. This is required for the upcoming "Fluid widgets" backport (#10568).
NOTE: This change is not backwards compatible, if you work with PHPs array_* functions on the query result. To work around this issue, you'll have to convert the query result to an array before by calling the QueryResult::toArray() method. We're planning to add a compatibility mode, but that's not yet implemented.

Resolves: #10566