[!!!][~TASK] Extbase (Configuration): Major rework of the ConfigurationManager
Configuration of controllers and actions is now stored in a global registry
($GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['extbase']['extensions']). But you
should never access this directly. Instead always retrieve the frameworkConfiguration
from the ConfigurationManager.
Inserting an Extbase plugin is now as simple as:
lib.foo = USER
lib.foo {
  userFunc = tx_extbase_core_bootstrap->run
  extensionName = YourExtension
  pluginName = YourPlugin
}

This is not really a breaking change as it does not change the public API. But it's not unlikely that it changes the behavior of your Extension in case you modified the TypoScript, that is generated by Tx_Extbase_Utility_Extension::configurePlugin().

NOTE: Unit tests of Extbase and Fluid v4 are broken currently. We'll be fixing those asap