[FEATURE] add option to select recursively in tcaTree
in some cases you might need or want to select a whole subtree - this is now possible.
Since ExtJs delivers nearly everything, just use it.

Add this to pending Documentation:
new Option in tcaTree[appearence]: allowRecursiveMode
If true, a toggle button in the toolbar is activated. If you toggle the button, a selection of a node, will automatically select/unselect all childnodes recursively, too.

Note: I added a Icon to the sprite, which has been regenerated via extdeveval (therefore the patch is more lengthy)

Change-Id: Ic79c2cddfeb111cfc2911bf6f4dfe51c6f36a14c
Resolves: #25061
Releases: 4.7