[BUGFIX] pagetree->expand contextmenu action does not work as expected
When I click "expand branch" in the new pagetree once it simply expands the next level und the page where I opened the context menu.

When I click the same context menu link (at the same page) a second time it expands all pages on all levels below the page where I clicked.

I'd expect the tree either doing the first OR the second action but not 2 different actions on the same useraction.

Change-Id: I4f6d2b5342e0d8c5f4e89d71b2527f8c640ad30e
Resolves: #M17394