[BUGFIX] pagetree->expand contextmenu action does not work as expected
When I click "expand branch" in the new pagetree once it simply expands the next level und the page where I opened the context menu.

When I click the same context menu link (at the same page) a second time it expands all pages on all levels below the page where I clicked.

I'd expect the tree either doing the first OR the second action but not 2 different actions on the same useraction.

Change-Id: I559156ba17b978f5d1c4a6859404d94efb007d1e
Resolves: #M17394
Reviewed-on: http://review.typo3.org/1047
Reviewed-by: Stefan Galinski <stefan.galinski@gmail.com>
Tested-by: Stefan Galinski <stefan.galinski@gmail.com>