[BUGFIX] No image generation with PHP-SAFE_MODE (GM/IM)
With PHP in SAFE_MODE all imagemagick/graphicsmagick calls will fail. To
support paths with whitespace characters quotes are added (using
escapeshellarg()). With safe_mode enabled PHP executes escapeshellcmd()
on the entire command. Now the quotes will be escaped and the command
will fail.
This patch changes that escapeshellarg is only if it is useful (i.e. if
there are whitespace characters or if characters were escaped).

Change-Id: I77d946e270c9400e0f4a3bfbcf1ca1741284c4d5
Resolves: #M17685