[BUGFIX] Deleted overlay missing in overlay priorities
The "deleted" overlay type is missing from the list of overlay
priorities as defined in
$TBE_STYLES['spriteIconApi']['spriteIconRecordOverlayPriorities'].
This patch fixes this by adding it as the first item in the list,
since one can assume "deleted" to be the most important status.

NOTE: it's hard to view the result of this patch as deleted records
are not normally shown in the TYPO3 BE. You may have deleted
records when working in a workspace, but they will still not be
visible after this patch due to a further bug in t3lib_iconworks.
See http://bugs.typo3.org/view.php?id=17117 for a discussion of that
further bug.

To view the effect of this patch, apply the patch from #M17117,
go to a workspace, delete an unversioned record and see the icon
change.

Change-Id: Ie2c1f6ef672c5eb73fbc37066fd7596380c64932
Resolves: #M17815
Reviewed-on: http://review.typo3.org/1089
Reviewed-by: Steffen Ritter <info@rs-websystems.de>
Tested-by: Steffen Ritter <info@rs-websystems.de>