[FEATURE] Make displayed fields in "Send to stage" layer configurable
Until now the recipients in the send to next stage popup are allways
checked per default and can be changed by the editor completely.

With this feature the administrator is able to
define the default state of the checkboxes. There is the choice between
"All are checked per default but it can be changed", "Some are checked
per default but can not be changed" and "All are checked per default
but con not be changed". Additionally it is possible to hide the
recipient checkboxes and the additional recipient field from the popup,
which is useful if there is a recipient configuration behind, see above.

Change-Id: Ib036d7dfabee8139bd7632184c9e46ce79d22979
Resolves: #13095