[BUGFIX] Media obj respect moveJsFromHeaderToFooter
If you set config.moveJsFromHeaderToFooter=1 in your TS template, and
add a Media object, the swfobject library is correctly added at the
bottom of the page, but the inline code calling the swfobject library
is added inline at the same place as the media object.
This results in the javascript error "swfobject is not defined" and
the media not appearing on the page.

This change makes use of t3lib_PageRenderer::addJsInlineCode to add
the inline code, so it respects config.moveJsFromHeaderToFooter=1.

Change-Id: I971ebb2c31568e40a3209dfcf7871010e830565e
Resolves: #25258
Reviewed-on: http://review.typo3.org/1062
Reviewed-by: Steffen Kamper
Reviewed-by: Björn Pedersen
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader
Releases: 4.6, 4.5
Reviewed-on: http://review.typo3.org/1547