[BUGFIX] Media obj respect moveJsFromHeaderToFooter
If you set config.moveJsFromHeaderToFooter=1 in your TS template, and
add a Media object, the swfobject library is correctly added at the
bottom of the page, but the inline code calling the swfobject library
is added inline at the same place as the media object.
This results in the javascript error "swfobject is not defined" and
the media not appearing on the page.

This change makes use of t3lib_PageRenderer::addJsInlineCode to add
the inline code, so it respects config.moveJsFromHeaderToFooter=1.

Change-Id: I0902ca60dab810dc58e7f1b46a07ad4ae6f1efda
Resolves: #M17868