[TASK] Refactor conditions in t3lib_extFileFunc
t3lib_extFileFunc has a lot of deeply nested conditions. These
conditions make understanding the code really hard.

Most of the else-branches only contain a log call, so it helps a lot to
invert the conditions, move the log call inside the if-branch and add a
break statement afterwards. The else branch can then be removed to
streamline the code.

Change-Id: Ic900fe3294142ebc89be4c3d420bb729a24d5452
Resolves: #24218
Reviewed-on: http://review.typo3.org/1087
Reviewed-by: Francois Suter
Tested-by: Francois Suter
Reviewed-by: Stefan Neufeind