[BUGFIX] EM update module doesn't accept cache_extension table
You see only the message:
To get and update the list, go to the section "Import extensions" and use the button "Retrieve/Update" even if table is filled.
The reason is that it looks for typo3temp/extensions.xml.hz instead checking if the table has records.

Solution:
Check table count instead.

Change-Id: Ide26c0fc5a3963f6bfa1a134ebfdbbf5f0b9b768
Resolves: #M17907