[BUGFIX] List module - Fix pointer issue in pagebrowser
When you have more than 50 records in the list module, you can use
the page browser to display the next set of records.
When using this on the pages table, the pointer to page 2,3,.. is
added to the link of each page record. Selecting such a link causes
the result listing of that page to also start from that specific result
page. If there are not enough records, none will be shown.

Resolves: #M11759

Change-Id: Ia112391533c40d4f2e56d41a6f893432480c8af1