[BUGFIX] BE list - Pointer issue on pagebrowser
When you have more than 50 records in the list module, you can use
the page browser to display the next set of records.
When using this on the pages table, the pointer is added to the link
of each page record, causing it to appear ithat no records or lesser
records are on that page when following that link.

Resolves: #M11759

Change-Id: Ia112391533c40d4f2e56d41a6f893432480c8af1