[FEATURE] t3lib_DB: Turn off pconnect by default
By default, TYPO3 uses *_pconnect() to connect to the database. This
is however only required in case connecting to the DB server is slow. In
other cases there is no advantage in turning this on, but there may be
disadvantages if the configuration of database and web server are not
aligned.
See also http://www.php.net/manual/en/features.persistent-connections.php

This commit was made based on a discussion on typo3-dev, see message
<mailman.12316.1298886389.648.typo3-dev@lists.typo3.org> and following.

Change-Id: Ibdc5f5c6045206039cbb23be47e1d2e648a94da5