[Bugfix] getUrl creates mailformed requests if no path is given
If you fetch an url like example.org without a '/' at the end,
you will get a 400 Bad request of only fetching headers (2).
If you get all content+headers (1) you will get a 404.
See http://forge.typo3.org/issues/13756.
All browsers I know automatically correct this to example.com/.
getUrl should be tolerant because users (editors) will not
understand the difference between example.org and example.org/.

This is a must have for linkvalidator, because otherwise we
will need to duplicate half of the getUrl function.
If the use of curl is enabled, this is automatically handled
by curl.

Solution: Check if a path is set, otherwise use '/' as path.

Change-Id: I86c7478d46d62f980735e5b80c6766e8bd811938
Resolves: #12225
Resolves: #13432
Resolves: #M16925