[BUGFIX] TCEFORM language selector: wrong relations
The language selector in the top of BE forms is intended to show (and
switch between) all translations of a certain record. All records in
non-default languages, that are not bound to a record in the default
language are shown as translations of each other.

The patch adds a destinction between translated records and records in
a non-default language to render the dropdown's labels and functions.

Change-Id: Iec7ee3dffbeab2e1fce45cd96c17ceffcc75f7aa
Resolves: 24958