[BUGFIX] TCEFORM language selector: wrong relations
The language selector in the top of BE forms is intended to show (and
switch between) all translations of a certain record. All records in
non-default languages, that are not bound to a record in the default
language are shown as translations of each other.
The patch adds a distinction between translated records and records in
a non-default language to render the dropdowns labels and functions.
Additionally fixes PHPDoc and indentation of SC_alt_doc::languageSwitch().

Change-Id: Iec7ee3dffbeab2e1fce45cd96c17ceffcc75f7aa
Resolves: #24958
Reviewed-on: http://review.typo3.org/1141
Reviewed-by: Andreas Wolf
Tested-by: Andreas Wolf