[FEATURE] Improve usability of group fields
If only one value is allowed in a group field (maxitems==1), the
present value will be removed when the user selects a new value via
element browser or file upload field.

If the list is disabled (by disable_controls setting) the user doesn't
have the possibility to remove items. To prevent a "write once" field
the first item in the list will be removed if the maxitems limit has
been reached.

Change-Id: Iac48371f672b78e7c2ea3450af581759986015bc
Resolves: #M14430