[FEATURE] EB is more uniform in the tree, upload & create folder form
This patch makes the element browser more conform in the three parts by
adding padding to the upload/create folder forms. And move the "display
thumbnails" checkbox to the top of the right tree navigation (right
below the bulk actions).

Furthermore, the three parts are visually splitted by a border which
made it necessary to add new id attributes to the both forms in order
to make them adressable in the stylesheets.

Resolves #M17942

Change-Id: I80211a8ace933fe4c44002a5eeb504bfa3360439