[FEATURE] Add support for FlexForm file reference
Now t3lib_befunc->getFlexFormDS supports pointers to a file, or to
a table containing the XML for the datastructure. I'm currently in
the need of storing a file location in this reference table, which
is not possible in current method.

My suggestion would be to validate the datastructure field in the
database to see if it's a file or XML, and handle it depending on
the outcome. Suggestion for a solution can be found in this changeset.

See the mantis bug report for a test extension to reproduce this.

Change-Id: Iae7b096b3b722dfc9eb2b59da10527e5f8febbf8
Resolves: #M17532