[BUGFIX] Fix missing patch for t3lib_compressor
For issue http://bugs.typo3.org/view.php?id=14303 a patch to
the t3lib_compressor has been submitted to trunk. Now the
error returned, and as it seems part of the patch has been
reverted since then.

Hereby a changeset to get this occurence in again.

Relates to: #M14303

Change-Id: Id394f9a67e6222f7663e719ddce602d5ada2d10c