[BUGFIX] Datetime picker returns only a date and no time value
When selecting a date with the datepicker, the time value is always 0:00.
The behaviour before 4.5 was that the timepicker return the selected date and the current time like 11:35 11-01-2011.
This patch adds a DateTimePicker for datetime fields that returns correct time and gives the possibility to change time value in the picker.

Change-Id: I2be56f5ecb56edd9c5619ae3bbf6d8cb8a091768
Resolves: #M17596