[BUGFIX] Abandon one time CSRF tokens
Abandon the extra security feature of having one time tokens and create tokens
which are valid during a whole login session. Additionally create only one random token,
store it in the session and create the real URL and form tokens by hashing the scope strings
with the secret session token.

To enable re-login, store the session token in the registry and retrieve it in case a
re-login happens.

Thanks to Marion Eher (Bluechip.at) for sponsoring
this fix with 75 beers during the bug auction at T3BOARD11.

Resolves: #25359
Change-Id: I078a6fa7f579026a33568fd0af114e5776c994da
Reviewed-on: http://review.typo3.org/1361
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel