[+BUGFIX] Avoiding cHash for nonCacheable actions prevent action from being called
With #8565 cHash is only appended for calls to cachable actions
This has the side effect, that calling non-cacheable actions
from a cached action on the same page fails to execute the
target action.
This fix reverts the changes from #8565 as we couldn't find
a better solution for now. See #13970 for details.

Change-Id: I0aebbca4a8df37ba7b7b5b63ceb0962990d814c7
Related: #13970