[BUGFIX] fix php fatal error of Upgrade wizard
The upgrade wizard of the install tool quits with a php fatal error if
php is configured to use registered globals.

When using registered globals, $GLOBALS['LANG'] may already be
initialized (with some string). The buggy line does not properly
initialize the language object if there is content in $GLOBALS['LANG'].
The solution changes this, so that $GLOBALS['LANG'] has to be an object
to prevent object creation. this line is also found in other places in
the typo3 source.

Change-Id: I1e553d233d6dac636d8a10a4ea5e60946e89d220
Resolves: #M17087