[BUGFIX] Fix PHP Fatal Error in Upgrade wizard
The upgrade wizard of the install tool quits with a php fatal error if
php is configured to use registered globals.

When using registered globals, $GLOBALS['LANG'] may already be
initialized (with some string). The buggy line does not properly
initialize the language object if there is content in $GLOBALS['LANG'].
The solution changes this, so that $GLOBALS['LANG'] has to be an object
to prevent object creation. this line is also found in other places in
the typo3 source.

Change-Id: I3eb45c77f21eae23ba2da99b3a461312c76475b2
Resolves: #25205