[BUGFIX] Fix PHP Catchable Fatal Error in t3lib_iconWorks
The hook in t3lib_iconWorks::mapRecordTypeToSpriteIconName() passes an
array to call a user function, but the method t3lib_div::callUserFunction
expects a variable being passed by reference.

Assigning the array to a variable fixes this error.

Thanks to Haymo Meran (Gentics) for sponsoring this fix with 80 beers
at the bug auction at T3BOARD11.

Change-Id: Ib0c0b2bd504fdbc346e8f70009bb60e82a66d8b0
Resolves: #25402
Reviewed-on: http://review.typo3.org/1378
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader