[BUGFIX] sys_action: Fix Record List page switching
If a list-records-task lists more than 100 records, the singe table page turning
functionality is shown on top of the table. If you want to switch the page (or
enter a page number) nothing happens or you get elsewhere than you are supposed
to be.

This happens due to wrong link generation. The Listmodule has been refactored, but the
sys_actions Module stayed the same. Link generation was broken.

This patch fixes the issues with sys_actions and list link generation by subclassing
the localRecordList class and adopting the link generation to sys_actions needs.

Change-Id: I6305901f0ede0efd0cf7305cbea1733bebce6a74
Resolves: #M15747