[BUGFIX] Resolves: #M15747; sys_action: Record list page turning not working
If a list-records-task lists more than 100 records, the singe table page turning
functionality is shown on top of the table. Unfortunately if you want to turn
the page (or enter a page number yourself) nothing happens or you get elsewhere in
newer TYPO3 versions.

This happens due to wrong link generation. The Listmodule has been refactored, but the
sys_actions Module stayed the same. Link generation was broken.

This patch fixes the issues with sys_actions and list link generation by subclassing
the localRecordList class and adopting the link generation to sys_actions needs.

Change-Id: I6305901f0ede0efd0cf7305cbea1733bebce6a74