[BUGFIX] Wrong display of csh icon and text in new record wizard
For each TCA-Table where you can add an element on a side, there is displayed
a csh icon with text (if text is configured). In db_new.php is a wrong
variable used so the csh text of the last tca-table is always used and for
each table. So it will also be displayed for tables which orginally got no csh text.

Additionally due to a missing break, the csh icon of a table is displayed
in the wrong line.

Change-Id: I7c4b0fec0e8c5c0a6947dec9c30b4c2ff3960d05
Resolves: #25371
Reviewed-on: http://review.typo3.org/1304
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog
Reviewed-by: Francois Suter
Tested-by: Francois Suter