[BUGFIX] wrong display of csh icon and text in new record wizzard
For each TCA-Table where you can add an element on a side, there is displayed
a csh icon with text if (text ist configured). In db_new.php is a wrong
variable used so the csh text of the last tca-table is allways used and for
each table. So it will also be displayed for tables which orginally got no csh text.

Additionally due to a missing break, csh icons of a tca-table is displayed
in the wrong line.

How to reproduce:
* install tt_news
* go to a site where you can insert tt_news records (not root)
* click the "create new record" icon.

Change-Id: I7c4b0fec0e8c5c0a6947dec9c30b4c2ff3960d05
Resolves: #25371