[BUGFIX] Wrong display of csh icon and text in new record wizard
For each TCA-Table where you can add an element on a side, there is displayed
a csh icon with text (if text is configured). In db_new.php is a wrong
variable used so the csh text of the last tca-table is always used and for
each table. So it will also be displayed for tables which orginally got no csh text.

Additionally due to a missing break, the csh icon of a table is displayed
in the wrong line.

Change-Id: I39cee0f65d40626eb2c893fbaae4a9f25ae6c7d5
Resolves: #25371