[BUGFIX] Restructure the random byte generator
Restructure the code to use the most performant methods first
if available. Take specialities of Windows OS and special
PHP versions into account.

Read/ generate more bytes than needed in one call, because it
does not cost (much) more to generate more random bytes, but it's
much cheaper for the next calls, because the bytes are already there.

Resolves: #23355
Releases: 4.6, 4.5, 4.4

Change-Id: I6bad300842f3da40c620b3d79b8116345a2749a0
Reviewed-on: http://review.typo3.org/4537
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter
Reviewed-by: Steffen Gebert
Tested-by: Steffen Gebert