[BUGFIX] Restructure the random byte generator
Restructure the code to use the most performant methods first
if available. Take specialities of Windows OS and special
PHP versions into account.

Read/ generate more bytes than needed in one call, because it
does not cost (much) more to generate more random bytes, but it's
much cheaper for the next calls, because the bytes are already there.

Resolves: #23355, #23860, #24410, #24440, #23496
Releases: 4.6, 4.5, 4.4, 4.3

Change-Id: I42eea55dcbcd8d8f5b1a6e9493993e9ccd967dfa
Reviewed-on: http://review.typo3.org/4555
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader
Reviewed-by: Tomasz Krawczyk
Tested-by: Tomasz Krawczyk
Reviewed-by: Steffen Gebert
Tested-by: Steffen Gebert