[BUGFIX] Installation from own TER not possible
It is possible to search the extension in my own TER using the extension
manager, but importing it fails. The reason is that the EM tries to fetch
the extension file from one of the orginal TYPO3 extension repository
mirrors, and not from the custom one.

Solution: adapt mirrorUrl fetching from tx_em_Connection_ExtDirectServer

Change-Id: Id50652b19b85f85c3156f031b53ad84fbd671de6
Resolves: #25387
Releases: 4.5,4.6