[BUGFIX] Installation from own TER not possible
It is possible to search the extension in my own TER using the extension manager, but importing it fails. The reason is that the EM tries to fetch the extension file from one of the orginal TYPO3 extension repository mirrors, and not from the custom one.

Solution: adapt mirrorUrl fetching from tx_em_Connection_ExtDirectServer

Change-Id: Ib699d26a6a2731df4eece71712244e067e85db03
Resolves: #25387