[FEATURE] Add helpful error messages in case of shortcut misconfiguration
This change improves the output of wrongly misconfigured shortcut pages with
helpful error messages. Instead of "The requested page didn't have a proper
connection to the tree-root!", hints are given dependent on the configured
shortcut mode and target.

Thanks to Volker Graubaum (e-netconsulting) for sponsoring this fix with 80 beers
during the Bug Auction at T3BOARD11.

Change-Id: I09d88f907fc459b10e1598e117ea03733be8e0d0
Resolves: #24223