[BUGFIX] unzip_path setting is used inconsistently
The setting $TYPO3_CONF_VARS']['BE']['unzip_path'] is used inconsistently in TYPO3.

In some cases, it expects the directory of the unzip binary. In other places, the
full path of the unzip binary is expected. This change adds a compatibilty layer
to cope with this.

Thanks to Tobias Liebig (networkteam GmbH) for sponsoring 60 beers during the
bug auction at T3BOARD11.

Change-Id: Ia34f8ae2ac89ed7c62fbb3165b4fda5388179273
Resolves: #25470