[BUGFIX] Default login (without EXT:felogin) form should be removed
When EXT:felogin is not installed, a FE login content element is available, that is
out of date and should not be used anymore.
As we ship felogin since 4.2 and install it by default, the default login form is
useless and confusing.

Thanks to naw.info for sponsoring 135 beers at the bug auction for this fix.

Change-Id: I765b3d391b06dac34550f46f19992c9cf9d1c5c7
Resolves: #22408