[FEATURE] Drop support for IE6-8 and ship Firefox Portable instead
Making the TYPO3 Backend work flawlessly with Internet Explorer below version 9
is a PITA. From the perspective of a better maintainability, I suggest to not
only drop the support for IE6, but also for versions 7 and 8.

Unfortunatley our target user group - the large enterprises - often have policies,
which disallow their staff to install good (read: Open Source) software and browsers
like Firefox. This change suggests to ship Firefox Portable with TYPO3 and offer it
as download at the login screen, if IE<9 is detected. This makes sure that also
such editors are able to work with the Backend, without requiring them to convince
their administrators to install free software.

Resolves: #20110401

Change-Id: Ia7fde479b037913c49955a185772814ab964edb8