[FEATURE] Add hook to override icon status overlay
Icons status overlays are defined in t3lib_iconWorks::
mapRecordOverlayToSpriteIconName(). However it is necessary
for "external" process (e.g. versioning) to override the
icon status. Currently a deleted record in a workspace will
not appear properly overlaid because only deleted = 1 is
checked and not also t3ver_state = 2. This hook makes it
possible for the version sysext to hook into the process and
set the right status.

Change-Id: I349012e402ab349ad810c57e46b0a7d5f8a46b03
Resolves: #25765
Releases: 4.6, 4.5