[BUGFIX] List module - Fix pointer issue in pagebrowser
When you have more than 50 records in the list module, you can use
the page browser to display the next set of records.
When using this on the pages table, the pointer to page 2,3,.. is
added to the link of each page record. Selecting such a link causes
the result listing of that page to also start from that specific result
page. If there are not enough records, none will be shown.

Change-Id: Iafa0c18b791c823a8dbd2357f85d7a28a0868871
Resolves: #20908
Reviewed-on: http://review.typo3.org/1538
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog