[+FEATURE] backport SignalSlotDispatcher
The SignalSlotDispatcher is a great way to have a centralized
EventHandler in Extbase. The SignalAspect was omitted as we don't
have AOP in Extbase. We have to dispatch all the Signals ourself.
Some exemplary Signals will be committed in another change.

The code of the SignalSlotDispatcher itself was backported
completely.

Change-Id: I7891ee06af495c73c6ec7883ee83fa4540d82350