[BUGFIX] Determine correct preview flag, if page is access restricted
If a page is access-restricted and the user is logged into the backend,
the page is displayed as a preview. To determine whether the page is
access restricted or not, it is not correct to only check whether
the page is hidden or not.

The patch implements a two level check. For the first level the page
is simply checked for some access-restriction. This is enough to set the
preview flag.

On the second level the pages in the rootline are checked for some
access restriction and if the flag 'extendToSubpages' is set, the
preview flag is also set.

Change-Id: I192830f05a8387218d72a48c972404cb267a9c0a
Resolves: #25894
Resolves: #16862
Resolves: #22259
Releases: 4.6, 4.5, 4.4, 4.3