[BUGFIX] TCEforms Suggest doesn't find static tables as user
As a Backend User, when using a suggest form next to a select
or group field containing entries from a static table I won't
get any results from this table. The Pid 0 is not accessible
from a users perspective, and thus checks against
t3lib_BEfunc::readPageAccess are unneccessary.

Added a condition in t3lib_TCEforms_Suggest_DefaultReceiver::
checkRecordAccess that checks for a is_static declaration in
TCA ctrl section of the affected table.

Change-Id: Id4a43f07c7d324b1c9c63a610db741b552df6f18
Fixes: #25938