[BUGFIX] Force icon overlay for deleted records
t3lib_iconWorks::mapRecordOverlayToSpriteIconName() defines the
overlays to display for any given record icon, but it not workspace-
aware. However a pending request adds a hook to override the
icon overlays. This enables the "version" extension to hook into
the process and force the overlays to "deleted" when the current
record's t3ver_state = 2 (i.e. it is deleted in a workspace).

Test in conjunction with https://review.typo3.org/#change,1459

Change-Id: I2c07e3ceae70aa971e1086dd1f902a960b862d09
Resolves: #24641
Releases: 4.6, 4.5