[BUGFIX] Removed dependency on TYPO3_MODE in saltedpasswords
Removed dependency on TYPO3_MODE constant in saltedpasswords which was
used to determine if the extension is enabled and which table (fe_users
or be_users) should be updated.

Since the mode is required in the init process of saltedpasswords where
the parent auth object is not available the new data in the info
array (introduced in #26410) is used.

Change-Id: I04180d283d6818b2e09d33f82036f6f88a5cb26c
Resolves: #22881