[BUGFIX] Fix clashing delimiter used by RTE Browse Links and FlexForms
The browse links class uses colon as a delimiter for different parameters
passed on tree updates. One of these parameters is the field name. When in
a context of a FlexForm colon is also used as delimiter in the field name,
thus creating a mismatch in the parameters in browse links.

To fix this the delimiter in browse links is changed to semicolon

Change-Id: I05dee76e5aaef50ade1d14f6988a46d8e254da1d
Resolves: #26583