[BUGFIX] phpTimeZone not correctly set from localconf.php
There's a bug in the PHP time zone implementation - set in
t3lib/config_default.php and overridden in localconf.php.

The timeZone variable used to set the default timezone takes the value
from TYPO3_CONF_VARS[phpTimeZone], which does not exist. It needs to be
corrected to TYPO3_CONF_VARS[SYS][phpTimeZone].

The effect is that the time zone for TYPO3 is always set to the server
time zone and is not set by the localconf.php / Install Tool value of
phpTimeZone.

This looks like it may have been an issue since phpTimeZone was added.

Thanks to Todd Hossack for the patch

Resolves: #26779
Releases: 4.6, 4.5
Change-Id: Ie443a7c1297156cb4c475fdcda38bafcbe20d716