[FEATURE] Make table labels for import static clickable
An extension that provides static data shows up in the extension manager
with a list of checkboxes for each table to import.
The table names are not clickable, one has to target the tiny checkbox.

This fix adds the missing <label> tags around the table names.

Thanks to Christian Weiske for providing this patch.

Change-Id: Icba748674a854cb3bda032cdfc797c4ad5584639
Resolves: #26926
Releases: 4.6