[TASK] Remove extension commerce specific code in TYPO3 Core
Class tslib_feUserAuth contains TER specific code. This issue is about
to remove this piece of code.

In specific this is EXT:commerce related code which has been
introduced with the TYPO3 Core Session Fixation Security Fix.

Whether fixed or not in commerce, Core should not fix code for
TER listed extensions. However, as commerce list currently very active
it should be possible for the contributors to fix problems if still
existing.

Change-Id: I2dae9c6f5fa24ce11f43ebd612523a00101e2990
Resolves: #27027
Reviewed-on: http://review.typo3.org/2364
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn