[BUGFIX] BE UI: Warning when creating multiple pages should be removed
When creating multiple new pages through the functions => wizard, and you'll create the pages you always get a JS question if you really want to do this. This is poor usability and stops the workflow because by that time, everyone should know what he's/she's doing. Additionally, this behaviour is very-"Windows"-like and is not really done in the same way on many other spots of the CMS.

Change-Id: I3990af110785f06b431ad1bdb8d93147397fbf4a
Resolves: #27104