[BUGFIX] Scheduler: use img title-tags for mouse-over
The 'Status Legend' in the Scheduler backend module don`t use
<img title="" /> tags. Only <img alt=""> tags.

This patch writes the value from alt=""-Tag to title="" tag.

This change is for 4.6 only, because the class in scheduler/mod1/index.php
was moved into scheduler/class.tx_scheduler_module.php in 4.6
To affect 4.5 and 4.4 with this change, this changes has to be made at
mod1/index.php in this releases (4.5, 4.4), too.

Change-Id: I401e3a129d8ef1dd7ea8c9d1130d1ed5ed1f6339
Resolves: #25661
Releases: 4.6