[BUGFIX] Scheduler: Task object state is not saved after execution
Every scheduler task object is saved as serialized object in the database.
If a schedular task run and modify internal member variables this modifications
are not saved in the database. On the next execution, all internal member variables
are back on default values.

This patch enables the possibility to enable the saving of the task object state
after the execution. Furthermore a new example task
"tx_scheduler_SaveTaskObjectStateTask" was added to show the new functionality.

Resolves: #25657
Releases: 4.6
Change-Id: I262584cdb4a896ea3b86be7c40fe184f2c55b724