[FEATURE] Link to Install Tool in User Settings
Enabling the Install Tool in the User Settings' Tab Admin Functions
does not provide a fast and useful way to lead one directly back to
the Install Tool. 

This feature provides the link directly under the Button that says
"Delete Install Tool Enable File" to easy switch back to the point
where the user wants to go.

Change-Id: I842d2148034d9e2cf3315621ee037744d7285d0b
Resolves: #27171