[BUGFIX] Scheduler: Make class tx_scheduler_Execution fully testable
The class tx_scheduler_Execution of scheduler is not fully testable,
because the class uses time()-function for calculations.
With the possibility to inject a custom timestamp it is possible to
write good unit tests for this class to save the functionality.

Thanks to Markus Krause who created the original patch.

Resolves: #25675
Releases: 4.6, 4.5, 4.4
Change-Id: I066c165cad985e5e99b46326b952f1affbe1072c