[BUGFIX] Do not suppress parse errors in ext_emconf.php files
In case of parse errors in an ext_emconf.php file, the error message
has been suppressed because of an @include(). Nevertheless, the EM
stops working with a blank page (without any error log entry).

Removing the @ shows the warning so that it easier for the user to
identify the erroneous extension.

Change-Id: I5dd67f1b486850ea1766dc8949b9541352632089
Resolves: #17506
Releases: 4.5, 4.6