[BUGFIX] TCEFORM language selector: wrong relations
The language selector in the top of BE forms is intended to show (and
switch between) all translations of a certain record. All records in
non-default languages, that are not bound to a record in the default
language are shown as translations of each other.
The patch adds a distinction between translated records and records in
a non-default language to render the dropdowns labels and functions.
Additionally fixes PHPDoc and indentation of
SC_alt_doc::languageSwitch().

Change-Id: I6292578a2e4b3807cb333ff85a6d6309b7d4c754
Resolves: #24958
Reviewed-on: http://review.typo3.org/2533
Reviewed-by: Thorsten Kahler
Tested-by: Thorsten Kahler
Reviewed-by: Andreas Wolf
Tested-by: Andreas Wolf