[FEATURE][!!!] Add XLIFF support (l10n)
First public version of the parser to support XLIFF + LLXML file as localization file format.

This patch is big, but it's not easy to have this kind of new feature in smaller patch.

This patch include:

- A parser factory to support multiple internationalization format (actually LLXML and XLIFF,
  parser for PHP array is not implemented yet). New parser can be added in the future.
- A new parser for XLIFF (supporting plurals, based on the parser of FLOW3)
- Add autoload for lang sysext
- Update the existing localization method (FE, BE, extbase) to support the new format
  of the LOCAL_LANG array
- New caching class based on the caching framework or file based
- Autodetect igbinary for cache un/serialization
- rtehtmlarea: update of the localization API

Currently the localization method work as before (only key based localization).

More functionality will be implemented in the new localization API (plurals, ...). The current parser
can be used on a TYPO3 instance with LLXML and/or XLIFF file. But if an extension use XLIFF the
language pack (typo3conf/l10n) must use the same format.

The parser and the format priority (by default: xlf, xml, php), can be configured in:

$TYPO3_CONF_VARS['SYS']['lang']

Warning: This patch break internationalization of many ExtJS based extension (like extension
manager, workspace, recycler ...). This bug print "Object" as label. I will push an other patch
with a new ExtJS API to handle it, and extension update to correct this. Some ExtJS extension are
broken (extension manager, recycler).

If you see "Object" as label in rtehtmlarea, please clear the cache of the RTE and
reload your backend.

Change-Id: I84d0a463f1b5448b4bca3f7d200315f397f4ea18