[BUGFIX] FE-Error "NO entry in the $TCA-array..."
Especially in installations with many domains, the following error
sporadically occured in Frontend:
> "NO entry in the $TCA-array for the table "[table]".
> This means that the function enableFields() is called with an invalid
> table name as argument."

The problem occured when a site is opened in frontend while a backend
user clears cache. It seems that at the beginning of the frontend
request the temp_*_ext_tables.php file still exists but when it should
be included it is got already deleted by the clear cache request. The
solution is to check if file exists before to include it, if not just
include all extensions ext_tables.php with stddb/load_ext_tables.php
"manually".

Change-Id: I4fe5e0256093df8190eb6f1a2754edf1e3aba405
Resolves: #22928
Releases: 4.6, 4.5, 4.4, 4.3