[BUGFIX] Wrong link if felogin reset password
When the reset password link for felogin is generated (to be sent in
the mail), config.absRefPrefix overrides
plugin.tx_felogin_pi1.feloginBaseURL, but feloginBaseURL is specifically
designed to be used when needing a different base URL. This patch
makes feloginBaseURL prevail.

Thanks to Stephan Bauer for the original patch.

Change-Id: I05e3e63a3ad4628e07396e2dd8f089cd18dfc7f9
Resolves: #22434
Releases: 4.6, 4.5