[BUGFIX] Use t3lib_cache_frontend_VariableFrontend as fallback frontend
According to wiki documentation, integrators are told to define both
'frontend' and 'backend' in their caching configuration.

This hardly makes sense as the frontend input format is the responsibility
of the extension developer, not the integrator and should not be forced
to be defined. Instead, it should fallback to the VariableFrontend if not
explicitly set.

Without fallback, t3lib_div::makeInstance() is called with an empty
class name which throws an exception when instantiated using reflection.

Change-Id: Iaf59864d86483e2d0a08b92c54dad579e2f81347
Resolves: #27677