[BUGFIX] Caching framework: Improve configuration handling
The cache manager takes care of correct cache configurations
given to the cache factory. The patch improves the handling of
default configuration for lazy cache setups.

Registration of new cache frontends or backends in TYPO3_CONF_VARS
is removed, new cache frontends or backends of extensions can be
defined by registering them to the autoloader only.

Change-Id: I62fd755b16d84dae2d13f0aab0062c95df2232c7
Resolves: #27689
Reviewed-on: http://review.typo3.org/2921
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Susanne Moog
Reviewed-by: Steffen Gebert
Tested-by: Steffen Gebert