[FEATURE] Enable closing of ExtJS FlashMessages
Some of the FlashMessages have suche a big space requirement that users
have to wait, until they disappear to continue working in the backend.

This patch makes the ExtJS FlashMessages disappear, if the user clicks
on close icon in the message header.

The icons a created by Fabien and added to the core css by Steffen.

Change-Id: I270dc0b6b4cfef945aa5802abcf586671db63e8b
Resolves: #27747
Releases: 4.6