[FEATURE] Implement extbase naming scheme in core autoloader
The patch enables the core autoloader to find class files that are named
according to the exbase class and file location naming scheme.
This renders ext_autoload.php entries obsolete for class locations
that stick to the extbase scheme even if extbase is not used
for an extension. The autoloaders within fluid and extbase are not
needed anymore.
Additionally, the class-to-file information is cached in the core
autoloader cache in order to reduce the number of file system lookups
for consecutive calls.
The patch is covered by unit tests to verify the complex logic.

Change-Id: I1af68a2d95cfc590eecffe6d06ea4e4eb58e11ec
Resolves: #21144
Related: #28249
Related: #28250
Reviewed-on: http://review.typo3.org/2985
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel