Language handling: splitLabels should be removed
The TYPO3 core currently handles three types of language labels:
1) splitlabels (very very old).
Based on the constant TYPO3_languages, a string is split by |. The only use was the loginLabels but this was changed in 4.3, so they only act as fallback right now.
2) locallang.php (very old)
An array for each language in a .php file.
3) locallang.xml (currently used)
An XML file with labels.
I urge to remove 1) and also disable 2) as we should clean up this mess of translation stuff.

Change-Id: Idd6905776af755cf54258fe6723ca1226816655d
Resolves: #26519