[BUGFIX] Fix fatal error: A cache with identifier "cache_hash" does not exist
When using the caching framework and calling an eID script containing locallang
labels, it may happen that data from cache_hash is tried to be accessed.

As there is no caching framework initialized at this time of execution,
a fatal error is triggered.

Fix the error by checking if the caching framework is initialized, otherwise
skip its usage.

Change-Id: Idafbcf0983fc189c5f71999a10c04ec4fd340f2b
Resolves: #25086