[BUGFIX] Fix exception error when using caching framework and eID scripts
When using the caching framework and calling an eID script containing
locallang labels, it may happen that data from cache_hash is tried to be
accessed.

As there is no caching framework initialized at this time of execution,
an exception warning is issued:
A cache with identifier "cache_hash" does not exist.

Fix the error by checking if the caching framework is initialized,
otherwise skip its usage.

Branches: 4.5
Change-Id: Idafbcf0983fc189c5f71999a10c04ec4fd340f2b
Resolves: #25086