[BUGFIX] Fix exception error when using caching framework and eID scripts
When using the caching framework and calling an eID script containing
locallang labels, it may happen that data from cache_hash is tried to be
accessed.

As there is no caching framework initialized at this time of execution,
an exception warning is issued:
A cache with identifier "cache_hash" does not exist.

Fix the error by checking if the caching framework is initialized,
otherwise skip its usage.

Change-Id: Iaeae13fe756b9b230f3eea919eaa3f045f2a6dcf
Branches: 4.5
Resolves: #25086
Reviewed-on: http://review.typo3.org/3068
Reviewed-by: Stefan Neufeind
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn