[FEATURE] Include HTTP Request2 for better HTTP handling
t3lib_div::getUrl() is known to be hudge and has a lot of problems
with edge cases. Especially it only implements a small subset of the
HTTP protocol. This is a problem for extensions which want to do some
more advanced HTTP magic (for example linkvalidator).

A discussion on v4 list suggested to implement pear/HTTP_Request2
and use t3lib_div::getUrl as a wrapper.

This is the first step which includes HTTP_Request2 and adds
t3lib_utility_httpRequest as wrapper which sets default options from
TYPO3_CONF_VARS.
Additionally some new TYPO3_CONF_VARS have been added and the old
cURL options have been deprecated.

Change-Id: I430685159cc966e13f1e833ca19ee8f0c75e400a
Resolves: #28344
Releases: 4.5,4.6