[BUGFIX] Deprecated missleading t3lib_div::cmpFQDN
Instead of using an IP address to compare the domain name,
the newly introduced method compareFQDN compares the domain
name against a list of domain names. The new method
compareFQDNByIP is backwards compatible and is being called
by the old method cmpFQDN. The old method cmpFQDN is set to
deprecated.

Change-Id: I71c0a9c09c54fb6f964a9e0385b38e19dadb194a
Resolves: #28046